added authorization on update and delete

+++ app/Http/Controllers/ListingController.php
@@ -68,6 +68,11 @@ class ListingController extends Controller
     }
 
     public function update(Request $request, Listing $listing) {
+
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -88,6 +93,11 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing) {
+        
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully!');
     }
